Add ListHeaderBar component and container to refactor structure
